refactor: consolidate V4.7 core models
