+++ custom_components/battery_smartflow_ai/core/__init__.py
@@ -1,0 +1,1 @@
+"""Platform-independent Battery SmartFlow AI core."""
